discovery-agent/requirements.txt
requests
beautifulsoup4
playwright
openpyxl
pandas
feedparser
schedule
pyyaml
